+++ src/axom/inlet/Container.hpp
@@ -310,7 +310,7 @@ std::vector<std::pair<std::string, std::string>> collectionIndicesWithPaths(
  *******************************************************************************
  */
 void updateUnexpectedNames(const std::string& accessedName,
-                           std::unordered_set<std::string>& unexpectedNames);
+                           std::vector<std::string>& unexpectedNames);
 
 }  // namespace detail
 
@@ -341,6 +341,8 @@ public:
    * \param [in] description Description of the Container
    * \param [in] reader Reference to the input file Reader class.
    * \param [in] sidreRootGroup Pointer to the already created Sidre Group.
+   * \param [in] unexpectedNames Reference to the global (relative to the Inlet
+   * hierarchy) list of unexpected names
    * \param [in] docEnabled Boolean indicating whether or not documentation
    * generation is enabled for input feck this Container instance belongs to.
    *****************************************************************************
@@ -349,12 +351,12 @@ public:
             const std::string& description,
             Reader& reader,
             axom::sidre::Group* sidreRootGroup,
-            std::unordered_set<std::string>& expected_names,
+            std::vector<std::string>& unexpectedNames,
             bool docEnabled = true)
     : m_name(name)
     , m_reader(reader)
     , m_sidreRootGroup(sidreRootGroup)
-    , m_unexpectedNames(expected_names)
+    , m_unexpectedNames(unexpectedNames)
     , m_docEnabled(docEnabled)
   {
     SLIC_ASSERT_MSG(m_sidreRootGroup != nullptr,
@@ -849,6 +851,21 @@ public:
 
   /*!
    *****************************************************************************
+   * \brief Set the strictness of this Container.
+   *
+   * Set whether this Container is strict, or not - i.e., whether entries other
+   * than those added to the schema should be allowed.
+   * The default behavior is to not be strict.
+   *
+   * \param [in] isStrict Boolean value of whether Container is strict
+   *
+   * \return Reference to this instance of Container
+   *****************************************************************************
+   */
+  Container& strict(bool isStrict = true);
+
+  /*!
+   *****************************************************************************
    * \brief Registers the function object that will verify this Container's contents
    * during the verification stage.
    * 
@@ -925,6 +942,16 @@ public:
    *****************************************************************************
    */
   std::string name() const;
+
+  /*!
+   *****************************************************************************
+   * \brief Returns the list of unexpected names "below" the calling container,
+   * i.e., entries in the input file structure (e.g., a Lua table or
+   * YAML dictionary) corresponding to the calling container that were not
+   * requested/retrieved via an add* call
+   *****************************************************************************
+   */
+  std::vector<std::string> unexpectedNames() const;
 
   /*!
    *****************************************************************************
@@ -1307,7 +1334,7 @@ private:
   axom::sidre::Group* m_sidreGroup;
   // Hold a reference to the global set of unexpected names so it can be updated when
   // things are added to this Container
-  std::unordered_set<std::string>& m_unexpectedNames;
+  std::vector<std::string>& m_unexpectedNames;
   bool m_docEnabled;
   std::unordered_map<std::string, std::unique_ptr<Container>> m_containerChildren;
   std::unordered_map<std::string, std::unique_ptr<Field>> m_fieldChildren;